feat: shrink-workstation-request-detail-test-override-002 - Split script, failure, and sparse fallback coverage into focused sibling tests
