+++ docs/internal/processes/development-guide-relevant-files.md
@@ -194,7 +194,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/src/App.current-selection.test.tsx` | Dedicated current-selection app regression surface | App-level selected-work, trace-card, responsive-layout, terminal-state, and mixed inference/script dispatch-history regressions should live here once a focused current-selection flow would otherwise outgrow the standard handwritten file or function size limits; when inference attempts or nested request/response bodies are collapsed by default, this seam must prove the app-level flow by expanding the relevant attempt and body disclosures before asserting payload or diagnostic text. |
 | `ui/src/App.import.test.tsx`, `ui/src/App.export-dialog.test.tsx`, `ui/src/App.export-submit.test.tsx`, `ui/src/App.layout-graph.test.tsx`, `ui/src/App.follow-up-submit.test.tsx`, `ui/src/App.follow-up-trace.test.tsx`, `ui/src/App.toolbar-locale.test.tsx`, and `ui/src/testing/app-shell-*.ts*` | Focused app-shell regression seams | When app-shell coverage grows past the repo size gates, split shell behaviors into focused app-level files by surface area, keep shared render, lifecycle, browser-stub, and default hook-mock builders local to `ui/src/testing/app-shell-*.ts*`, route shared `useCurrentFactoryDocument` setup through that harness instead of per-file duplicate mocks, and keep the harness responsible for the app-level `/factory-sessions` bootstrap fetch now that the shell mounts session tabs before the dashboard body. Keep only feature-specific seeding in the owning `App.*.test.tsx` file, and prefer named local scenario fixtures over new wrapper helpers when that seeding repeats. Prefer keeping both the focused files and their local helpers under the standard handwritten size limit before adding any new explicit Biome allowlist entries. |
 | `ui/src/features/current-selection/workstation-request-detail.tsx` | Workstation-request detail fallback seam | Request-detail rendering, script-versus-inference alternate panels, sparse projected field fallbacks, and the omission of low-value request-count summary rows should stay centralized here so current-selection request diagnostics keep one observable detail contract instead of splitting fallback copy across summary cards or app shells. |
-| `ui/src/features/current-selection/workstation-request-detail.test.tsx` | Workstation-request detail regression surface | Customer-visible request-detail regressions should assert rendered markdown/plain-text prompt handling, hidden request-count summary labels, plus sparse request, trace, script-request, and script-response fallback copy here rather than relying on helper-only tests or broader app smoke coverage. |
+| `ui/src/features/current-selection/components/workstation-request-detail.*.test.tsx` | Workstation-request detail regression surface | Customer-visible workstation-request detail regressions should stay split by observable behavior into focused sibling files for inference, script attempts, sparse fallbacks, failed outcomes, and consumed-work selection; keep the assertions on rendered regions, buttons, and localized copy rather than helper-only fixtures or broader app smoke coverage. |
 | `ui/src/features/current-selection/inference-attempt.tsx` and `ui/src/features/current-selection/execution-details.test.tsx` | Inference-attempt disclosure seam | Current-selection inference attempts should own their own collapsed-by-default disclosure state here, keep attempt status and concise timing visible in the collapsed summary, and prove per-attempt `aria-expanded` behavior plus independent expansion through the focused execution-details regression surface instead of broader widget smoke tests. |
 | `ui/src/features/current-selection/detail-card-shared.tsx` and `ui/src/features/current-selection/request-authored-text.test.tsx` | Shared authored-body formatting seam | Markdown-safe authored body rendering for current-selection detail cards should stay centralized here through the neutral `AuthoredBodyText` formatter, while request-specific wrappers and tests continue proving headings, lists, code, plain-text fallback, and inert raw-HTML handling without duplicating formatter logic in request or inference card components. |
 | `ui/src/features/current-selection/detail-card-shared.tsx`, `provider-session-attempts.tsx`, `provider-session-detail-panel.tsx`, and `selected-work-dispatch-history-card.tsx` | Current-selection emphasis color seam | Shared current-selection action/link foreground emphasis belongs in `detail-card-shared.tsx`, while current-dispatch card, provider-session transcript/status treatments, and badge emphasis belong in the feature owners that consume those shared constants; use semantic combinations such as `border-af-accent-border bg-af-accent-surface text-af-text` for selected/current emphasis and `border-af-border bg-af-surface-raised text-af-text-muted` for neutral status pills or code panels instead of reintroducing `on-foreground` or slash-opacity recipes. |
